fix: async validation not canceled
When async validation is started and something changes so validation is
called again, the previous validator is not canceled.

Changed async validation timer to depend on check that uses resolvable
promise to check if async validation should occur, so the promise can be
resolved with false by subsequent calls to cancel previous validations.

Closes #2